Revert "Moved components to a new campaign folder"
