+++ components/campaign-territory-list.tsx
