+++ components/campaign.tsx
@@ -3,7 +3,7 @@
 import { useEffect, useState } from 'react';
 import { createClient } from "@/utils/supabase/client";
 import Tabs from '@/components/tabs';
-import TerritoryList from '@/components/campaign/campaign-territory-list';
+import TerritoryList from '@/components/campaign-territory-list';
 import { Button } from '@/components/ui/button';
 import Modal from '@/components/modal';
 import { useToast } from "@/components/ui/use-toast";
